Move SummarizeDiff into TOCSummaries.
